backend/requirements.txt
django
djangorestframework
faker
django-debug-toolbar
django-redis
